extracts data transforming into own functions and improves code readability
